chore: remove unnecessary activity pages & components
