--- bciers/apps/reporting/src/app/bceidbusiness/industry_user/reports/[version_id]/facilities/[facility_id]/activities/pulp_and_paper_production/page.tsx
@@ -1,24 +1,0 @@
-import { actionHandler } from "@bciers/actions";
-import PulpAndPaperProduction from "@reporting/src/app/components/activities/pulpAndPaperProduction";
-import { Suspense } from "react";
-
-export default async function Page() {
-  const reportDate = "2024-04-01"; // This should be passed in once we have a path to this page from starting a report
-  const activityData = await actionHandler(
-    `reporting/initial-activity-data?activity_name=Pulp and paper production&report_date=${reportDate}`,
-    "GET",
-    "",
-  );
-  const activityDataObject = JSON.parse(activityData);
-
-  return (
-    <>
-      <Suspense fallback="Loading Schema">
-        <PulpAndPaperProduction
-          activityData={activityDataObject}
-          reportDate={reportDate}
-        />
-      </Suspense>
-    </>
-  );
-}
