--- bciers/apps/reporting/src/app/bceidbusiness/industry_user_admin/reports/[version_id]/facilities/[facility_id]/activities/gsc_other_than_non_compression/page.tsx
@@ -1,23 +1,0 @@
-import { Suspense } from "react";
-import { actionHandler } from "@bciers/actions";
-import safeJsonParse from "libs/utils/safeJsonParse";
-import GSCOtherThanNonCompression from "@reporting/src/app/components/activities/gSCOtherThanNonCompression";
-
-export default async function Page() {
-  const reportDate = "2024-04-01"; // This should be passed in once we have a path to this page from starting a report
-  const activityData = await actionHandler(
-    `reporting/initial-activity-data?activity_name=General stationary combustion, other than non-compression and non-processing combustion&report_date=${reportDate}`,
-    "GET",
-    "",
-  );
-  const activityDataObject = safeJsonParse(activityData);
-
-  return (
-    <Suspense fallback="Loading Schema">
-      <GSCOtherThanNonCompression
-        activityData={activityDataObject}
-        reportDate={reportDate}
-      />
-    </Suspense>
-  );
-}
